refactor: eliminate duplicate code
